+++ Sources/PayPalUI/PayPalButton.swift
@@ -2,10 +2,7 @@ import UIKit
 import SwiftUI
 
 /// Configuration for PayPal button
-public final class PayPalButton: PaymentButton, UIViewRepresentable {
-
-    /// SwiftUI button action
-    var action: () -> Void = { }
+public final class PayPalButton: PaymentButton {
 
     /// Available colors for PayPalButton.
     public enum Color: String {
@@ -48,8 +45,7 @@ public final class PayPalButton: PaymentButton, UIViewRepresentable {
         color: Color = .gold,
         edges: PaymentButtonEdges = .softEdges,
         size: PaymentButtonSize = .collapsed,
-        label: Label? = nil,
-        _ action: @escaping () -> Void = { }
+        label: Label? = nil
     ) {
         self.init(
             fundingSource: .payPal,
@@ -59,29 +55,80 @@ public final class PayPalButton: PaymentButton, UIViewRepresentable {
             insets: insets,
             label: label?.label
         )
-        self.action = action
     }
 
     deinit {}
+}
 
-    // MARK: - UIViewRepresentable methods
+public extension PayPalButton {
+    
+    /// PayPalButton for SwiftUI
+    struct Representable: UIViewRepresentable {
+        
+        private var action: () -> Void = { }
+    
+        private let button: PayPalButton
+        
+        /// Initialize a PayPalButton
+        /// - Parameters:
+        ///   - insets: Edge insets of the button, defining the spacing of the button's edges relative to its content.
+        ///   - color: Color of the button. Default to gold if not provided.
+        ///   - edges: Edges of the button. Default to softEdges if not provided.
+        ///   - size: Size of the button. Default to collapsed if not provided.
+        ///   - label: Label displayed next to the button's logo. Default to no label.
+        public init(
+            insets: NSDirectionalEdgeInsets? = nil,
+            color: PayPalButton.Color = .gold,
+            edges: PaymentButtonEdges = .softEdges,
+            size: PaymentButtonSize = .collapsed,
+            label: PayPalButton.Label? = nil,
+            _ action: @escaping () -> Void = { }
+        ) {
+            button = PayPalButton(
+                fundingSource: .payPal,
+                color: color.color,
+                edges: edges,
+                size: size,
+                insets: insets,
+                label: label?.label
+            )
+            self.action = action
+        }
+       
+        // MARK: - UIViewRepresentable methods
+        // TODO: Make unit test for UIVRepresentable methods: https://engineering.paypalcorp.com/jira/browse/DTNOR-623
 
-    public func makeCoordinator() -> Coordinator {
-        Coordinator(action: action)
+        public func makeCoordinator() -> Coordinator {
+            Coordinator(action: action)
+        }
+
+        public func makeUIView(context: Context) -> UIView {
+            let view = UIView()
+
+            view.addSubview(button)
+            button.addTarget(context.coordinator, action: #selector(Coordinator.onAction(_:)), for: .touchUpInside)
+
+            return view
+        }
+
+        public func updateUIView(_ uiView: UIView, context: Context) {
+            context.coordinator.action = action
+        }
     }
+}
 
-    public func makeUIView(context: Context) -> UIView {
-        let view = UIView()
-        let payPalButton = self
+// MARK: PayPalButton Preview
 
-        view.addSubview(payPalButton)
+struct PayPalButtonView: View {
 
-        payPalButton.addTarget(context.coordinator, action: #selector(Coordinator.onAction(_:)), for: .touchUpInside)
-
-        return view
+    var body: some View {
+        PayPalButton.Representable()
     }
+}
 
-    public func updateUIView(_ uiView: UIView, context: Context) {
-        context.coordinator.action = action
+struct PayPalButtonView_Preview: PreviewProvider {
+
+    static var previews: some View {
+        PayPalButtonView()
     }
 }