+++ Sources/PayPalUI/PayPalCreditButton.swift
@@ -2,10 +2,7 @@ import UIKit
 import SwiftUI
 
 /// Configuration for PayPal Credit button
-public final class PayPalCreditButton: PaymentButton, UIViewRepresentable {
-
-    /// SwiftUI button action
-    var action: () -> Void = { }
+public final class PayPalCreditButton: PaymentButton {
 
     /**
     Available colors for PayPalCreditButton.
@@ -30,8 +27,7 @@ public final class PayPalCreditButton: PaymentButton, UIViewRepresentable {
         insets: NSDirectionalEdgeInsets? = nil,
         color: Color = .darkBlue,
         edges: PaymentButtonEdges = .softEdges,
-        size: PaymentButtonSize = .collapsed,
-        _ action: @escaping () -> Void = { }
+        size: PaymentButtonSize = .collapsed
     ) {
         self.init(
             fundingSource: PaymentButtonFundingSource.credit,
@@ -41,28 +37,77 @@ public final class PayPalCreditButton: PaymentButton, UIViewRepresentable {
             insets: insets,
             label: nil
         )
-        self.action = action
     }
 
     deinit {}
+}
 
-    // MARK: - UIViewRepresentable methods
+public extension PayPalCreditButton {
+    
+    /// PayPalCreditButton for SwiftUI
+    struct Representable: UIViewRepresentable {
+        
+        private let button: PayPalCreditButton
+        private var action: () -> Void = { }
+            
+        /// Initialize a PayPalCreditButton
+        /// - Parameters:
+        ///   - insets: Edge insets of the button, defining the spacing of the button's edges relative to its content.
+        ///   - color: Color of the button. Default to dark blue if not provided.
+        ///   - edges: Edges of the button. Default to softEdges if not provided.
+        ///   - size: Size of the button. Default to collapsed if not provided.
+        public init(
+            insets: NSDirectionalEdgeInsets? = nil,
+            color: PayPalCreditButton.Color = .darkBlue,
+            edges: PaymentButtonEdges = .softEdges,
+            size: PaymentButtonSize = .collapsed,
+            _ action: @escaping () -> Void = { }
+        ) {
+            self.button = PayPalCreditButton(
+                fundingSource: PaymentButtonFundingSource.credit,
+                color: color.color,
+                edges: edges,
+                size: size,
+                insets: insets,
+                label: nil
+            )
+            self.action = action
+        }
+        
+        // MARK: - UIViewRepresentable methods
+        // TODO: Make unit test for UIVRepresentable methods: https://engineering.paypalcorp.com/jira/browse/DTNOR-623
 
-    public func makeCoordinator() -> Coordinator {
-        Coordinator(action: action)
+        public func makeCoordinator() -> Coordinator {
+            Coordinator(action: action)
+        }
+
+        public func makeUIView(context: Context) -> UIView {
+            let view = UIView()
+
+            view.addSubview(button)
+            button.addTarget(context.coordinator, action: #selector(Coordinator.onAction(_:)), for: .touchUpInside)
+
+            return view
+        }
+
+        public func updateUIView(_ uiView: UIView, context: Context) {
+            context.coordinator.action = action
+        }
     }
+}
 
-    public func makeUIView(context: Context) -> UIView {
-        let view = UIView()
-        let payPalCreditButton = self
+// MARK: PayPalCreditButton Preview
 
-        view.addSubview(payPalCreditButton)
-        payPalCreditButton.addTarget(context.coordinator, action: #selector(Coordinator.onAction(_:)), for: .touchUpInside)
+struct PayPalCreditButtonView: View {
 
-        return view
+    var body: some View {
+        PayPalCreditButton.Representable()
     }
+}
 
-    public func updateUIView(_ uiView: UIView, context: Context) {
-        context.coordinator.action = action
+struct PayPalCreditButtonView_Preview: PreviewProvider {
+
+    static var previews: some View {
+        PayPalCreditButtonView()
     }
 }