+++ UnitTests/PayPalUI_Tests/PayPalButton_Tests.swift
@@ -3,23 +3,36 @@ import XCTest
 
 class PayPalButton_Tests: XCTestCase {
 
+    // MARK: - PayPalButton for UIKit
+    
     func testInit_whenPayPalButtonCreated_hasUIImageFromAssets() {
-        let payPalButton = PayPalButton()
-        XCTAssertEqual(payPalButton.imageView?.image, UIImage(named: "PayPalLogo"))
+        let sut = PayPalButton()
+        XCTAssertEqual(sut.imageView?.image, UIImage(named: "PayPalLogo"))
     }
 
-    func testInit_whenPayPalButtonCreated_hasUIColorFromAssets() {
-        let payPalButton = PayPalButton()
-        XCTAssertEqual(payPalButton.containerView.backgroundColor, PaymentButtonColor.gold.color)
+    func testInit_whenPayPalButtonCreated_hasDefaultUIValuess() {
+        let sut = PayPalButton()
+        XCTAssertEqual(sut.edges, PaymentButtonEdges.softEdges)
+        XCTAssertEqual(sut.size, PaymentButtonSize.collapsed)
+        XCTAssertEqual(sut.color, PaymentButtonColor.gold)
+        XCTAssertNil(sut.label)
+        XCTAssertNil(sut.insets)
     }
 
-    func testInit_whenSwiftUIPayPalButtonCreated_canInit() {
-        let action = { }
-        let payPalButton = PayPalButton { }
-        let coordinator = Coordinator(action: action)
+    // MARK: - PayPalButton.Representable for SwiftUI
+    
+    func testMakeCoordinator_whenOnActionIsCalled_executesActionPassedInInitializer() {
+        let expectation = expectation(description: "Action is called")
+        let sut = PayPalButton.Representable {
+            expectation.fulfill()
+        }
+        let coordinator = sut.makeCoordinator()
 
-        XCTAssertNotNil(payPalButton)
-        XCTAssertNotNil(payPalButton.makeCoordinator())
-        XCTAssertNotNil(coordinator.onAction(action))
+        coordinator.onAction(self)
+        waitForExpectations(timeout: 1) { error  in
+            if error != nil {
+                XCTFail("Action passed in PayPalButton.Representable was never called.")
+            }
+        }
     }
 }